- Added unit tests for FindNearbyGyms - Added unit tests for AddGymToFavorites - Added unit tests for DislikeGym - Added unit tests for ObserveCurrentLocation

+++ app/src/main/java/com/insearching/urbansports/core/util/LocationUtils.kt
@@ -29,9 +29,9 @@ object LocationUtils {
         startLng: Double,
         endLat: Double,
         endLng: Double
-    ): Float {
+    ): Double {
         val results = FloatArray(1)
         Location.distanceBetween(startLat, startLng, endLat, endLng, results)
-        return results[0]
+        return results[0].toDouble()
     }
 }
